legacy/pyapp-1.3/requirements.txt
PySide6>=6.9.3
pyinstaller>=6.0.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
pyqtdarktheme>=0.1.7